Refactor test utils to a separate dev crate and enable version 1 tests without validation.
