+++ Cargo.toml
@@ -6,15 +6,18 @@ edition = "2021"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [features]
-# When enabled, rust the SBML syntactic test suite as part of unit tests.
+# When enabled, run the SBML syntactic test suite as part of unit tests.
 # This is mainly used for the purpose of code coverage computation.
 sbml_test_suite = []
 
 [dependencies]
 const_format = "0.2.31"
-macros = { path = "macros" }
 phf = { version = "0.11.2", features = ["macros"] }
-strum = "0.25.0"
-strum_macros = "0.25"
-xml-doc = { git = "https://github.com/daemontus/xml-doc" }
+strum = "0.26"
+strum_macros = "0.26"
 regex = "1.10.3"
+xml-doc = { git = "https://github.com/daemontus/xml-doc" }
+sbml-macros = { path = "sbml-macros" }
+
+[dev-dependencies]
+sbml-test-suite = { path = "sbml-test-suite" }
