+++ sbml-macros/README.md
@@ -2,4 +2,4 @@
 
 This crate provides useful macros for generating boilerplate code 
 for the main `lib-sbml` project. It is not intended to be published
-or used independently on the main project.
+or used independently of the main project.
