+++ src/core/unit_definition.rs
@@ -1,7 +1,7 @@
 use crate::core::sbase::SbmlUtils;
 use crate::core::Unit;
 use crate::xml::{OptionalChild, XmlDefault, XmlDocument, XmlElement, XmlList};
-use macros::{SBase, XmlWrapper};
+use sbml_macros::{SBase, XmlWrapper};
 
 /// Individual unit definition
 #[derive(Clone, Debug, XmlWrapper, SBase)]
