+++ src/core/validation/mod.rs
@@ -28,8 +28,6 @@ mod parameter;
 mod reaction;
 mod rule;
 mod species;
-#[cfg(test)]
-mod test_suite;
 mod unit;
 mod unit_definition;
 mod xml;
